backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
pywebpush>=2.0.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
cryptography>=42.0.0
